Performance issue in constructing subject class.
ExecutorService should be shared between different instances of the ToplogicalOrthographicProjectionSceneBuffer. This should be thread-safe. There is no reason or purpose to providing each their own thread pool, and the pools are expensive to construct. Given that the scene buffers are reconstructed each frame, there is a significant performance penalty if the pool is not shared and static.

Jeremy. Allen. Wildsmith.